requirements.txt
fitz
numpy
openai
opencv_python
Pillow
pymupdf
